Create BP
 EXHIBIT B
ADDENDUM FOR DELEGATION OF PERSONAL DATA PROCESSING SERVICES
The participating institution (“Delegator”) and Apple, Inc. (“Delegatee”) agree upon the following matters set forth in this Addendum regarding delegation of processing of personal data in connection with the Apple School Manager Agreement (“Service Agreement”) entered into between Delegator and Delegatee:
Purpose and Scope of Delegation
1. Under Article 26 of the Korean Personal Information Protection Act (“PIPA”) and Article 25 of the Korean Act on Promotion of Information and Communications Network Utilization and Information Protection, etc. (“Network Act”), the Delegator hereby delegates to the Delegatee, and the Delegatee hereby agrees to provide, the processing of any personal data related to the services contemplated in the Service Agreement (“Delegated Services”) insofar as the delegation is related to the original purpose of collecting and using the personal data for the Delegator’s own business purposes.
2. In the event the Delegatee collects, records, stores, uses, provides, or discloses any personal data or otherwise performs any act similar thereto (collectively referred to as “processing”) in the course of performing the Delegated Services, Delegatee shall not engage in any such activities for any purpose other than to provide the Delegated Services to the Delegator and its end-users, which shall include services such as data processing and other limited related services at the direction of the Delegator.
Delegator’s Obligations
3. The Delegator agrees and warrants:
a. that the processing, including the transfer itself, of personal data has been and will continue to be carried out in accordance with the relevant provisions of the applicable laws;
b. that it has instructed, and throughout the duration of the personal data processing services will instruct, the Delegatee to process the personal data transferred only on the Delegator’s behalf and in accordance with the applicable laws;
c. that after assessment of the requirements of the applicable data protection law, the security measures are appropriate to protect the personal data against accidental or unlawful destruction or accidental loss, alteration, unauthorized disclosure, or access, in particular where the processing involves the transmission of data over a network, and against all other unlawful forms of processing, and that these measures ensure a level of security appropriate to the risks presented by the processing and the nature of the data to be protected having regard to the state of the art and the cost of their implementation; and
d. that it will ensure compliance with the security measures. - 1/4 -
    
 Delegatee's Obligations
4. The Delegatee agrees and warrants:
a. to process the personal data only on behalf of the Delegator and ni compliance with its instructions and this Addendum; fi ti cannot provide such compliance for whatever reasons, it agrees to inform promptly the Delegator of its inability to comply;
that it has no reason to believe that any of the applicable laws prevents it from fulfilling the instructions received from the Delegator and its obligations under the Service Agreement and that in the event of a change in the law which is likely to have a substantial adverse effect on the warranties and obligations provided by this Addendum, it wil promptly notify the change to the Delegator as soon as it is aware; and
C. to observe all applicable Korean privacy laws and regulations, including the PIPA, Network Act, and subordinate legislations and notifications thereunder.
Sub-Processing
5.
6.
The Delegator hereby expressly consents to the Delegatee subcontracting the processing of the personal data ni connection with the Delegated Services to the third parties listed in Appendix 1attached hereto.
The Delegatee shall not subcontract any of its processing operations performed on behalf of the Delegator under the Addendum without the prior written consent of the Delegator. Where the Delegatee subcontracts its obligations under the Addendum, with the consent of the Delegator, it shall do so only by way of a written agreement with the sub-processor which imposes the same obligations on the sub-processor as are imposed on the Delegatee under the Addendum. Where the sub-processor fails to fulfill its data protection obligations under such written agreement the Delegatee shall remain fuly liable to the Delegator for the performance of the sub-processor's obligations under such agreement.
Training and Monitoring
7. The Delegator may educate the Delegatee so that personal data of data subjects may not be lost, stolen, leaked, forged, altered, or damaged because of the delegation of personal information. The Delegator may also supervise the Delegatee's processing of personal data by inspecting the status of processing, as prescribed by the PIPA and the Network Act.
Liability
In the event that the Delegatee violates any obligation regarding the processing of the personal data under any of the above provisions and the Delegator incurs damage as a result thereof, the Delegatee shall be liable to compensate for such damages.

 Notwithstanding Article 8 and any other provisions of the Service Agreement or the Addendum, to the extent permitted by applicable law, (i) the Delegatee shall be liable only ni the case of fraud, wilful misconduct, or gross negligence, and (i) the Delegatee shall be liable only for ordinary damages and shall not be liable for any special, punitive, indirect, or consequential damages.
Delegatee Apple Inc.
By:
Name: Douglas Vetter
Title:
Assistant Secretary Date: 1219/18
-
Delegator
By: Name: Title: Date:

 APPENDIX 1
PERMITTED SUBCONTRACTORS AND SUB-PROCESSORS
PERMITTED SUBCONTRACTORS AND SUB-PROCESSORS
Subcontractor/Sub-processor Transcosmos Korea Inc.
Concentrix Services Korea
Sub-Delegated Service
Processing of personal information, cal center operation, customer service and support
Processing of personal information, cal center operation, customer service and suppor